fixed cart functionality

+++ public/js/cart.js
@@ -1,45 +1,72 @@
-const decreaseBtn = document.querySelector('#decrease');
-const increaseBtn = document.querySelector('#increase');
-const delBtn = document.querySelector("#delete");
-const qtyField = document.querySelector("#qty");
-const cartId = document.querySelector("#cartId");
+const decreaseBtn = document.querySelector('.decrease');
+const increaseBtn = document.querySelector('.increase');
+const delBtn = document.querySelector(".delete");
+const cartId = document.querySelector(".cartId");
 
 const decreaseQty = async (event) => {
-    event.preventDefault();
-
+    const id = event.target.getAttribute("data-id");
+    const qtyField = document.getElementById(`qty-${id}`);
     let qty = parseInt(qtyField.textContent);
-    if (qty > 0) {
-        qty -= 1;
-        qtyField.textContent = qty;
-        const response = await fetch(`/api/cart/${cartId.textContent}`, {
-            method: 'PUT',
-            body: JSON.stringify({ qty }),
-            headers: { 'Content-Type': 'application/json' },
-        });
+
+    if (qty > 1) {
+        if (event.target.hasAttribute('data-id')) {
+            
+            qty -= 1;
+            qtyField.textContent = qty;
+
+            const response = await fetch(`/api/cart/${id}`, {
+              method: 'PUT',
+              body: JSON.stringify({ qty }),
+              headers: { 'Content-Type': 'application/json' },
+            });
+
+            if (response.ok) {
+              document.location.reload();
+            } else {
+              alert('Failed to change qty');
+            }
+        }
     } 
 };
 
 const increaseQty = async (event) => {
-    event.preventDefault();
-
+    const id = event.target.getAttribute("data-id");
+    const qtyField = document.getElementById(`qty-${id}`)
     let qty = parseInt(qtyField.textContent);
-    if (qty < 100) {
+
+    if (event.target.hasAttribute('data-id')) {
+    
         qty += 1;
         qtyField.textContent = qty;
-        const response = await fetch(`/api/cart/${cartId.textContent}`, {
+
+        const response = await fetch(`/api/cart/${id}`, {
             method: 'PUT',
             body: JSON.stringify({ qty }),
             headers: { 'Content-Type': 'application/json' },
         });
+
+        if (response.ok) {
+            document.location.reload();
+        } else {
+            alert('Failed to change qty');
+        }
     } 
 };
 
 const delBtnHandler = async (event) => {
-    event.preventDefault();
+    if (event.target.hasAttribute('data-id')) {
+        const id = event.target.getAttribute('data-id');
+    
+        const response = await fetch(`/api/cart/${id}`, {
+          method: 'DELETE',
+        });
 
-    const response = await fetch(`/api/cart/${cartId.textContent}`, {
-        method: "DELETE"
-    });
+        if (response.ok) {
+          document.location.reload();
+        } else {
+          alert('Failed to delete product');
+        }
+    }
 };
 
 decreaseBtn.addEventListener("click", decreaseQty);
